Rendering System working quite good (changed startTime in Time.java from float to long as it had issues with FPS precision)

+++ src/main/java/util/Time.java
@@ -1,7 +1,7 @@
 package util;
 
 public class Time {
-    public static float timeStarted = System.nanoTime();
+    public static long timeStarted = System.nanoTime();
 
     public static float getTime(){
         return ((float) ((System.nanoTime() - timeStarted)*(1e-9)));
